HWUMC-101 Fix errant group-leave-not-allowed error
When user A was editing a user who was in groups that A is not a manager
for, the UI previously didn't send a value for the disabled checkbox.
Therefore the server assumed that the user had unchecked it, and tripped
the security tests. Now it's forced to be sent a true value from the UI,
while still permitting overly pokey users to get a nasty error if they
choose to modify the raw HTML of the form.

Hopefully this is the last of Magz' issues.